requirements.txt
streamlit>=1.27.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
protobuf>=4.25.1
typing-extensions>=4.5.0
